fix player visiblity on first move, fix duration settings in editor timeline objects (no decimals)
implemented by GKGamerRU

+++ Assets/Scripts/Level/LvlEditor/OSBEditorObject.cs
@@ -269,11 +269,9 @@ public class OSBEditorObject : MonoBehaviour
 
     public void DragActiveZone()
     {
-
-
         Vector2 point;
         RectTransformUtility.ScreenPointToLocalPointInRectangle(activeZone, Input.mousePosition, null, out point);
 
-        assignedActor.objParams["Duration"].number.expression = ((point.x - activeZoneOffset) * 10f).ToString();
+        assignedActor.objParams["Duration"].number.expression = ((int)((point.x - activeZoneOffset) * 10f)).ToString();
     }
 }
